+++ components/log-viewer-webui/client/src/components/ResultsTimeline/typings.ts
@@ -1,0 +1,17 @@
+import dayjs from "dayjs";
+import Duration from "dayjs/plugin/duration";
+
+
+interface TimelineConfig {
+    range: {begin: dayjs.Dayjs; end: dayjs.Dayjs};
+    bucketDuration: Duration.Duration;
+}
+
+
+interface ChartTooltipItemRaw {
+    x: number;
+}
+
+export type {
+    ChartTooltipItemRaw, TimelineConfig,
+};
