+++ components/log-viewer-webui/client/src/pages/SearchPage/SearchState.tsx
@@ -13,14 +13,14 @@ import {
  */
 const SEARCH_STATE_DEFAULT = Object.freeze({
     queryString: "",
-    selectedOption: DEFAULT_TIME_RANGE,
+    timeRangeOption: DEFAULT_TIME_RANGE,
     timeRange: TIME_RANGE_OPTION_DAYJS_MAP[DEFAULT_TIME_RANGE],
 });
 
 interface SearchState {
     queryString: string;
-    selectedOption: TIME_RANGE_OPTION;
-    setSelectedOption: (option: TIME_RANGE_OPTION) => void;
+    timeRangeOption: TIME_RANGE_OPTION;
+    setTimeRangeOption: (option: TIME_RANGE_OPTION) => void;
     timeRange: [dayjs.Dayjs, dayjs.Dayjs];
     updateQueryString: (query: string) => void;
     updateTimeRange: (range: [dayjs.Dayjs, dayjs.Dayjs]) => void;
@@ -28,8 +28,8 @@ interface SearchState {
 
 const useSearchStore = create<SearchState>((set) => ({
     ...SEARCH_STATE_DEFAULT,
-    setSelectedOption: (option: TIME_RANGE_OPTION) => {
-        set({selectedOption: option});
+    setTimeRangeOption: (option: TIME_RANGE_OPTION) => {
+        set({timeRangeOption: option});
     },
     updateQueryString: (query) => {
         set({queryString: query});
